agent_test_backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
agno
openai
trafilatura
sqlalchemy